[Feat/#45] 스케줄 API (#50)
* [Feat/#42] ui 변경사항 반영

* [Feat/#45] schedule service 구현

* [Feat/#45] schedule datasource 구현

* [Feat/#45] schedule repository 구현

* [Feat/#45] schedule api 연동

* [Feat/#45] weekly sticker api 구현

* [Feat/#45] weekly sticker api 연동

* [Fix/#45] weekly sticker api 엔드포인트 수정

* [Fix/#45] weekly sticker 넘길 때 깜박임 이슈 해결
- loading 중 데이터 활용

* [Fix/#45] week offset 실패 시 원복 로직 추가

* [Fix/#45] ScheduleUiModel dDay 타입 Long으로 변경

+++ app/src/main/java/com/swyp/firsttodo/presentation/todo/component/ScheduleList.kt
@@ -36,7 +36,7 @@ import com.swyp.firsttodo.presentation.common.component.task.TaskEditPopup
 
 data class ScheduleUiModel(
     val scheduleId: Long,
-    val dDay: Int,
+    val dDay: Long,
     val title: String,
     val date: String,
     val rawDate: String,
